refactor(rust): prefer using `program` explicitly on `OxcAst`
