+++ crates/rolldown_oxc_utils/src/oxc_ast/helpers.rs
@@ -8,7 +8,7 @@ use crate::OxcAst;
 
 impl OxcAst {
   pub fn is_body_empty(&self) -> bool {
-    self.inner.with_dependent(|_, program| program.body.is_empty())
+    self.program.with_dependent(|_, program| program.body.is_empty())
   }
 
   pub fn make_semantic<'ast>(
@@ -21,7 +21,7 @@ impl OxcAst {
   }
 
   pub fn make_symbol_table_and_scope_tree(&self) -> (SymbolTable, ScopeTree) {
-    self.inner.with_dependent(|dep, program| {
+    self.program.with_dependent(|dep, program| {
       let semantic = Self::make_semantic(&dep.source, program, self.source_type);
       semantic.into_symbol_table_and_scope_tree()
     })
